pyproject.toml
[project]
name = "pdf-text-extraction"
version = "0.1.0"
description = "Extract text from scanned PDFs using OCR and save to markdown"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "PyMuPDF>=1.23.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]

[tool.uv]
dev-dependencies = []

